Rewrite in typescript
